+++ src/main/kotlin/it/czerwinski/intellij/wavefront/editor/ui/MaterialPropertiesTableCellRenderer.kt
@@ -21,6 +21,7 @@ import com.intellij.util.ui.JBEmptyBorder
 import it.czerwinski.intellij.common.ui.ColorLabel
 import it.czerwinski.intellij.wavefront.lang.psi.MtlIlluminationValueElement
 import it.czerwinski.intellij.wavefront.lang.psi.MtlReflectionType
+import it.czerwinski.intellij.wavefront.lang.psi.MtlScalarChannel
 import java.awt.Color
 import java.awt.Component
 import javax.swing.JTable
@@ -41,6 +42,7 @@ class MaterialPropertiesTableCellRenderer : TableCellRenderer {
         is Float -> JBLabel(value.toString(), SwingConstants.TRAILING)
         is Color -> ColorLabel(value)
         is MtlIlluminationValueElement.Illumination -> JBLabel(value.description)
+        is MtlScalarChannel -> JBLabel(value.description)
         is MtlReflectionType -> JBLabel(value.description)
         else -> JBLabel(value?.toString().orEmpty())
     }.apply {