+++ src/main/kotlin/it/czerwinski/intellij/wavefront/lang/psi/MtlElementFactory.kt
@@ -78,9 +78,16 @@ object MtlElementFactory {
             .getChildrenOfType<MtlValueModifierOption>().single()
     }
 
+    fun createScalarChannelOption(project: Project, value: MtlScalarChannel): MtlScalarChannelOption {
+        val file = createFile(project, text = "newmtl temp\ndisp -imfchan ${value.optionValue} filename.png")
+        return (file.firstChild as MtlMaterial)
+            .getChildrenOfType<MtlScalarTextureElement>().single()
+            .getChildrenOfType<MtlScalarChannelOption>().single()
+    }
+
     fun createReflectionTextureElement(
         project: Project,
-        reflectionType: MtlReflectionType,
+        reflectionType: MtlReflectionType = MtlReflectionType.SPHERE,
         filename: String
     ): MtlReflectionTextureElement {
         val file = createFile(project, text = "newmtl temp\nrefl -type ${reflectionType.name.lowercase()} $filename")
@@ -89,7 +96,7 @@ object MtlElementFactory {
 
     fun createReflectionTypeOption(
         project: Project,
-        reflectionType: MtlReflectionType
+        reflectionType: MtlReflectionType = MtlReflectionType.SPHERE
     ): MtlReflectionTypeOption {
         return createReflectionTextureElement(project, reflectionType, filename = "filename.png")
             .getChildrenOfType<MtlReflectionTypeOption>().single()